+++ tests/components/hue/test_light.py
@@ -1,13 +1,9 @@
 """Philips Hue lights platform tests."""
 import asyncio
-from collections import deque
 import logging
 from unittest.mock import Mock
 
 import aiohue
-from aiohue.groups import Groups
-from aiohue.lights import Lights
-import pytest
 
 from homeassistant import config_entries
 from homeassistant.components import hue
@@ -173,48 +169,6 @@ LIGHT_GAMUT = color.GamutType(
     color.XYPoint(0.138, 0.08),
 )
 LIGHT_GAMUT_TYPE = "A"
-
-
-@pytest.fixture
-def mock_bridge(hass):
-    """Mock a Hue bridge."""
-    bridge = Mock(
-        hass=hass,
-        available=True,
-        authorized=True,
-        allow_unreachable=False,
-        allow_groups=False,
-        api=Mock(),
-        reset_jobs=[],
-        spec=hue.HueBridge,
-    )
-    bridge.mock_requests = []
-    # We're using a deque so we can schedule multiple responses
-    # and also means that `popleft()` will blow up if we get more updates
-    # than expected.
-    bridge.mock_light_responses = deque()
-    bridge.mock_group_responses = deque()
-
-    async def mock_request(method, path, **kwargs):
-        kwargs["method"] = method
-        kwargs["path"] = path
-        bridge.mock_requests.append(kwargs)
-
-        if path == "lights":
-            return bridge.mock_light_responses.popleft()
-        if path == "groups":
-            return bridge.mock_group_responses.popleft()
-        return None
-
-    async def async_request_call(task):
-        await task()
-
-    bridge.async_request_call = async_request_call
-    bridge.api.config.apiversion = "9.9.9"
-    bridge.api.lights = Lights({}, mock_request)
-    bridge.api.groups = Groups({}, mock_request)
-
-    return bridge
 
 
 async def setup_bridge(hass, mock_bridge):
